Change verification script to run by digest
