+++ templates/env.template
@@ -1,5 +1,6 @@
 LOG_GROUP_NAME={{ DOMAIN_NAME }}
 DOCKSTORE_VERSION={{ DOCKSTORE_VERSION }}
+DOCKSTORE_IMAGE_DIGEST={{ DOCKSTORE_IMAGE_DIGEST }}
 GITHUB_APP_PRIVATE_KEY_FILE={{ GITHUB_APP_PRIVATE_KEY_FILE }}
 GALAXY_PLUGIN_VERSION={{ GALAXY_PLUGIN_VERSION }}
 DATABASE_DOMAIN={{ DATABASE_DOMAIN }}
